backend/requirements.txt
basketball-reference-scraper==1.0.31
beautifulsoup4==4.8.2
blinker==1.7.0
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==2.0.12
click==8.1.7
flask==3.0.2
Flask-Cors==4.0.0
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==4.6.5
MarkupSafe==2.1.5
numpy
packaging==23.2
pandas==1.3.1
python-dateutil==2.8.1
pytz==2019.3
requests
six==1.13.0
soupsieve==1.9.5
Unidecode==1.2.0
urllib3==1.26.18
werkzeug==3.0.1
zipp==3.17.0
newsapi-python==0.2.7
python-dotenv==1.0.1